Wave 7: 106 functions matched, 332 -> 438, and two harness bugs found
Eight agents, 106 functions assigned, 106 matched and promoted. Both builds
reproduce SHA1 14dd0b22c894865867aff89e8116b2dffae25605.

The batch-wide compile_probe held up again: several agents cleared 12-15
functions with zero wasted try_match calls. Repeated families paid best --
five 44-byte siblings at 0x0803C8F0 were one template with the member
changed, and six identical 32-byte accessors at 0x08018CE8 really are
byte-identical in the ROM.

Two bugs in my own tooling, both found by chasing a single "failure":

trymatch.py rejected sub_08011B34 at 39 of 40 bytes. The difference was not
real. asm/ is disassembled output, so a pool word holding an address gets
symbolized as whichever symbol sits there -- gUnknown_03000040 -- while the
candidate's `base + 16` emits the same address as gUnknown_03000000 with the
difference in the inline addend. Both link to identical bytes, confirmed by
building the ROM from the C. trymatch now resolves symbol+addend through the
ELF symbol table before calling a mismatch, and is strict about it: same
offsets, same types, ABS32 only, both symbols known, and every differing byte
inside a relocation site. A deliberately wrong bound (15 instead of 16) is
still rejected.

index_functions.py could not see a function returning a function pointer --
`void (*f(void))(void)` nests the name inside parentheses. sub_080366DC and
sub_080366E8 were promoted into the ROM and still counted as asm, so the work
queue kept offering them.

Progress: 438 of 4,659 functions (9.4%), 12,632 bytes.

+++ include/unknown-functions.h
@@ -3,14 +3,20 @@
 
 #include "global.h"
 
-void sub_08011550(s16);
-void sub_0801156C(s16);
-void sub_08011588(s16);
-void sub_080115B4(s16);
-void sub_080115E0(s16, ProcPtr);
-void sub_080115F8(s16, ProcPtr);
-void sub_08011610(s16, ProcPtr);
-void sub_0801163C(s16, ProcPtr);
+/* These eight take a 32-bit first argument, not the s16 they were declared with
+ * until wave 7. All eight open with a bare `adds r4, r0, #0` and no extension,
+ * and the ROM's own ARM PROMOTE_MODE zero-extends any sub-word parameter that
+ * has to survive a call -- `s16` compiles to `adds r4, r0, #0; lsl #16; lsr #16`
+ * and cannot match. Their callers in proc.c pass `dataImm`, which is s16, so
+ * the caller side is unaffected either way. */
+void sub_08011550(int);
+void sub_0801156C(int);
+void sub_08011588(int);
+void sub_080115B4(int);
+void sub_080115E0(int, ProcPtr);
+void sub_080115F8(int, ProcPtr);
+void sub_08011610(int, ProcPtr);
+void sub_0801163C(int, ProcPtr);
 void Decompress(u8 *, void *);
 
 void sub_08012358(void);
